photos/requirements.txt
an-example-pypi-project==0.0.0
Django==2.2.13
django-cors-headers==2.4.0
django-extensions==2.1.3
django-filter==2.0.0
django-simple-history==2.5.1
djangorestframework==3.9.1
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2018.9
six==1.12.0
djangorestframework-simplejwt-4.4.0
pyjwt-1.7.1
